Конечный вариант 7 спринт
