+++ MovieQuiz/Services/MovieQuizViewControllerProtocol.swift
@@ -10,11 +10,8 @@ import Foundation
 protocol MovieQuizViewControllerProtocol: AnyObject {
     func show(quiz step: QuizStepViewModel)
     func show(quiz result: QuizResultsViewModel)
-    
     func highlightImageBorder(isCorrectAnswer: Bool)
-        
     func showLoadingIndicator()
     func hideLoadingIndicator()
-    
     func showNetworkError(message: String)
 }
