+++ MovieQuiz/Services/MoviesLoader.swift
@@ -7,21 +7,23 @@
 
 import Foundation
 
-//Протокол для загрузки фильмов
 protocol MoviesLoading {
     func loadMovies(handler: @escaping (Result<MostPopularMovies, Error>) -> Void)
 }
-
 struct MoviesLoader: MoviesLoading {
-    private let networkClient = NetworkClient()
     
+    private let networkClient: NetworkRouting
+    init(networkClient: NetworkRouting = NetworkClient()) {
+        self.networkClient = networkClient
+    }
+
     private var mostPopularMoviesUrl: URL {
+        // Если мы не смогли преобразовать строку в URL, то приложение упадёт с ошибкой
         guard let url = URL(string: "https://imdb-api.com/en/API/Top250Movies/k_zcuw1ytf") else {
             preconditionFailure("Unable to construct mostPopularMoviesUrl")
         }
         return url
     }
-    
     func loadMovies(handler: @escaping (Result<MostPopularMovies, Error>) -> Void) {
         networkClient.fetch(url: mostPopularMoviesUrl) { result in
             switch result {
@@ -38,4 +40,5 @@ struct MoviesLoader: MoviesLoading {
         }
     }
 }
+
