+++ MovieQuiz/Services/NetworkClient.swift
@@ -12,7 +12,7 @@ protocol NetworkRouting {
 }
 
 // Отвечает за загрузку данных по URL
-struct NetworkClient {
+struct NetworkClient: NetworkRouting {
 
     private enum NetworkError: Error {
         case codeError
